Refactor the nav url helpers so they can be shared
